+++ src/components/ProgressBar/ProgressBar.story.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { storiesOf } from '@storybook/react';
-import { withKnobs, text, number } from '@storybook/addon-knobs';
+import { withKnobs, text, number, boolean } from '@storybook/addon-knobs';
 import { withInfo } from '@storybook/addon-info';
 
 import ProgressBar from '../ProgressBar';
@@ -16,11 +16,24 @@ stories
   ));
 
 stories
-  .add('states of pb', withInfo()(() => {
+  .add('Progress with knobs', withInfo()(() => {
+    return (
+      <ProgressBar
+        displayError={ boolean('displayError', false) }
+        max={ number('max', 100) }
+        min={ number('min', 0) }
+        progress={ number('progress', 60) }
+        topLabel={ text('topLabel', 'hola esto es un label') }
+        leftLabel={ text('leftLabel', 'Allocated traffic') }
+        rightLabel={ text('rightLabel', 'Available traffic') }
+      />);
+  }))
+  .add('Progress bad news', withInfo()(() => {
     return (
       <ProgressBar
         max={ number('max', 100) }
         min={ number('min', 0) }
+        displayError={ true }
         progress={ number('progress', 60) }
         topLabel={ text('topLabel', 'hola esto es un label') }
         leftLabel={ text('leftLabel', 'Allocated traffic') }