+++ src/main.js
@@ -20,6 +20,7 @@ export { default as HelpPopover } from './components/HelpPopover';
 export { default as Input } from './components/Input';
 export { default as Label } from './components/Label';
 export { default as Link } from './components/Link';
+export { default as ListGroup } from './components/ListGroup';
 export { default as OverlayWrapper } from './components/OverlayWrapper';
 export { default as Pagination } from './components/Pagination';
 export { default as ProgressBar } from './components/ProgressBar';